requirements.txt
Django==1.9.11
django-markdown-app==0.8.6
docopt==0.6.2
Markdown==2.6.7
Pygments==2.1.3
requests==2.11.1
yarg==0.1.9
